requirement.txt
asgiref==3.7.2
Django==4.2.6
django-appconf==1.0.5
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-imagekit==5.0.0
django-js-asset==2.1.0
mysqlclient==2.2.0
pilkit==3.0
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
